Add commitment option to getAccountInfo
When sending a NFT, it is normally necessary to create an ATA (associated token account) for the receivers wallet. Only then the SPL token (the NFT) can be send.
To do this in one step, we have to get the newly created account info with the commitment of "confirmed", otherwise the sending of the token will fail.